refactor and docstring
